backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

mysql-connector-python==8.2.0

google-generativeai==0.3.2
python-dotenv==1.0.0

face-recognition==1.3.0
face-recognition-models==0.3.0

opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

numpy==1.24.3
Pillow==10.1.0
requests==2.31.0

gtts==2.4.0
SpeechRecognition==3.10.0
pydub==0.25.1

fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.1.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
python-multipart==0.0.6
pydantic==2.4.2